backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
boto3>=1.34.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
shortuuid>=1.0.0
polars>=0.20.0
pandera>=0.18.0
openpyxl>=3.1.0
fastapi_redis_cache>=0.0.1
redis>=4.5.0
requests>=2.31.0
pydruid>=0.6.0
strawberry-graphql>=0.200.0
httpx>=0.24.0
numpy>=1.24.0